feat(frontend): core component for all wallet loaders
