+++ src/frontend/src/lib/components/core/Loaders.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
-	import BtcLoaderWallets from '$btc/components/core/BtcLoaderWallets.svelte';
-	import IcLoaderWallets from '$icp/components/core/IcLoaderWallets.svelte';
 	import LoaderBalances from '$icp-eth/components/core/LoaderBalances.svelte';
 	import Loader from '$lib/components/core/Loader.svelte';
 	import LoaderMetamask from '$lib/components/core/LoaderMetamask.svelte';
+	import LoaderWallets from '$lib/components/core/LoaderWallets.svelte';
 	import ExchangeWorker from '$lib/components/exchange/ExchangeWorker.svelte';
 	import AddressGuard from '$lib/components/guard/AddressGuard.svelte';
 </script>
@@ -11,15 +10,13 @@
 <AddressGuard>
 	<Loader>
 		<LoaderBalances>
-			<IcLoaderWallets>
-				<BtcLoaderWallets>
-					<ExchangeWorker>
-						<LoaderMetamask>
-							<slot />
-						</LoaderMetamask>
-					</ExchangeWorker>
-				</BtcLoaderWallets>
-			</IcLoaderWallets>
+			<LoaderWallets>
+				<ExchangeWorker>
+					<LoaderMetamask>
+						<slot />
+					</LoaderMetamask>
+				</ExchangeWorker>
+			</LoaderWallets>
 		</LoaderBalances>
 	</Loader>
 </AddressGuard>
